fix: improve examples deps
